# sdk/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Open Configuration Management API",
    "version": "v1",
    "description": "Multi-tenant configuration management with versioning and access control"
  },
  "paths": {
...
    "/api/v1/~cast-formats/": {
      "get": {
        "operationId": "list_cast_formats",
        "summary": "List Cast Formats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CastFormatsListSchema"
                }
              }
            }
          }
        },
        "description": "Return all REST-supported cast formats and their option schemas.",
        "tags": [
          "Resolve"
        ],
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
    "/api/v1/ns/{namespace}/~resolve/{path}/~download/{token}": {
      "get": {
        "operationId": "download_resolved_artifact",
        "summary": "Download Resolved Artifact",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "path",
            "schema": {
              "title": "Path",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "token",
            "schema": {
              "title": "Token",
              "type": "string"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "description": "Download a previously resolved artifact using a signed token.\n\nRequest authentication is not used; the short-lived token in the URL is the\nsole credential (signed-URL semantics). Any Authorization header is ignored.",
        "tags": [
          "Resolve"
        ]
      }
    },
    "/api/v1/ns/{namespace}/~resolve/{path}": {
      "get": {
        "operationId": "resolve_config",
        "summary": "Resolve Config",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "path",
            "schema": {
              "title": "Path",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "version",
            "schema": {
              "default": "latest",
              "title": "Version",
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "cast",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cast"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "trace_only",
            "schema": {
              "default": false,
              "title": "Trace Only",
              "type": "boolean"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "mark-stable",
            "schema": {
              "default": false,
              "title": "Mark-Stable",
              "type": "boolean"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "ignore-configs-with-missing-tags",
            "schema": {
              "default": false,
              "title": "Ignore-Configs-With-Missing-Tags",
              "type": "boolean"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "no-creds",
            "schema": {
              "default": false,
              "title": "No-Creds",
              "type": "boolean"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponseSchema"
                }
              }
            }
          }
        },
        "description": "Resolve a Config (or every Config under a folder path) through the\nparameters → extend → render → cast pipeline.\n\nWhen ``mark-stable=true``, advance the reserved ``stable`` tag on resolved\nConfig root(s) to the version that was just resolved (Configs only).\n\nWhen ``ignore-configs-with-missing-tags=true`` and ``path`` is a folder,\nskip Configs that do not have the requested ``version`` tag or number\ninstead of failing the whole folder resolve.\n\nWhen ``no-creds=true``, secret parameters are replaced with the dummy\nvalue ``<secret-value-placeholder>`` without fetching or decrypting secrets\n(``secret:resolve`` is not required).\n\nWhen authenticated as a Resolver, ``path`` is relative to the resolver's\nscope. Use ``'.'`` to resolve the scope root folder.",
        "tags": [
          "Resolve"
        ],
        "security": [
          {
            "ResolverAuth": []
          },
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
    "/api/v1/ns/{namespace}/~resolve-parameters/{path}": {
      "get": {
        "operationId": "resolve_parameters",
        "summary": "Resolve Parameters",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "path",
            "schema": {
              "title": "Path",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "version",
            "schema": {
              "default": "latest",
              "title": "Version",
              "type": "string"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "no-creds",
            "schema": {
              "default": false,
              "title": "No-Creds",
              "type": "boolean"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveParametersResponseSchema"
                }
              }
            }
          }
        },
        "description": "Return effective parameter values for a single Config (debug).\n\nWhen ``no-creds=true``, secret parameters show the dummy value\n``<secret-value-placeholder>`` without fetching secrets (``secret:resolve``\nis not required).\n\nWhen authenticated as a Resolver, ``path`` is relative to the resolver's\nscope. Use ``'.'`` to refer to the scope root (must be a Config, not a\nfolder — a folder path is rejected by the underlying resolver).",
        "tags": [
          "Resolve"
        ],
        "security": [
          {
            "ResolverAuth": []
          },
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
    "/api/v1/ns/{namespace}/~resolve-draft/{path}": {
      "post": {
        "operationId": "resolve_draft_config",
        "summary": "Resolve Draft Config",
        "parameters": [
          {
            "in": "path",
            "name": "namespace",
            "schema": {
              "title": "Namespace",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "path",
            "name": "path",
            "schema": {
              "title": "Path",
              "type": "string"
            },
            "required": true
          },
          {
            "in": "query",
            "name": "cast",
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cast"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "trace_only",
            "schema": {
              "default": false,
              "title": "Trace Only",
              "type": "boolean"
            },
            "required": false
          },
          {
            "in": "query",
            "name": "no-creds",
            "schema": {
              "default": false,
              "title": "No-Creds",
              "type": "boolean"
            },
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResolveResponseSchema"
                }
              }
            }
          }
        },
        "description": "Resolve unsaved draft YAML at the given path without persisting it.\n\nThe full parameters → extend → render → cast pipeline runs against the\nsubmitted content. ``config:resolve`` is required on the path and all\ntransitive participants. Results are returned as download URLs (same\nshape as ``GET /~resolve/{path}``). Response items use ``version=0``.",
        "tags": [
          "Resolve"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfigDocument",
                "description": "Config YAML document.",
                "format": "textarea",
                "title": "ConfigDocument",
                "type": "string"
              },
              "example": {
                "foo": "bar",
                "baz": 42
              }
            },
            "application/yaml": {
              "schema": {
                "description": "Config YAML document.",
                "format": "textarea",
                "title": "ConfigDocument",
                "type": "string"
              },
              "example": "foo: bar\nbaz: 42"
            },
            "application/octet-stream": {
              "title": "File upload",
              "schema": {
                "type": "string",
                "format": "binary"
              },
              "description": "Upload config as a file",
              "example": "config.yaml"
            }
          },
          "required": true,
          "description": "Document as raw body: pick a text Content-Type to paste, or application/octet-stream to upload a file"
        },
        "security": [
          {
            "OAuth2AuthorizationCodeBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfigDocument": {
        "description": "Config YAML document.",
        "format": "textarea",
        "title": "ConfigDocument",
        "type": "string"
      },
...
      "CastFormatSchema": {
        "description": "One supported cast output format and its option JSON Schema.",
        "properties": {
          "format": {
            "description": "Cast format identifier (yaml, json, env, …)",
            "title": "Format",
            "type": "string"
          },
          "options_schema": {
            "additionalProperties": true,
            "description": "JSON Schema for ``cast_option_*`` query parameters",
            "title": "Options Schema",
            "type": "object"
          }
        },
        "required": [
          "format",
          "options_schema"
        ],
        "title": "CastFormatSchema",
        "type": "object"
      },
      "CastFormatsListSchema": {
        "properties": {
          "formats": {
            "items": {
              "$ref": "#/components/schemas/CastFormatSchema"
            },
            "title": "Formats",
            "type": "array"
          }
        },
        "required": [
          "formats"
        ],
        "title": "CastFormatsListSchema",
        "type": "object"
      },
      "EffectiveResolverHooksSchema": {
        "description": "Hook commands from the effective resolver configuration.",
        "properties": {
          "validate": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validate"
          },
          "validate_all": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Validate All"
          },
          "post_resolve": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Resolve"
          },
          "post_resolve_all": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Post Resolve All"
          }
        },
        "title": "EffectiveResolverHooksSchema",
        "type": "object"
      },
      "EffectiveResolverSchema": {
        "description": "Effective resolver configuration for a resolver-authenticated resolve call.",
        "properties": {
          "cast": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cast"
          },
          "parameters": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parameters"
          },
          "hooks": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EffectiveResolverHooksSchema"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "title": "EffectiveResolverSchema",
        "type": "object"
      },
      "ResolveResponseSchema": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ResolvedItemSchema"
            },
            "title": "Items",
            "type": "array"
          },
          "length": {
            "title": "Length",
            "type": "integer"
          },
          "trace_only": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Trace Only"
          },
          "root": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Root"
          },
          "resolver": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/EffectiveResolverSchema"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "items",
          "length"
        ],
        "title": "ResolveResponseSchema",
        "type": "object"
      },
      "ResolvedItemSchema": {
        "description": "One resolved output document.\n\nMirrors the design's resolve response item: `url` is the short-lived\nsigned download URL for this artifact; the resolved bytes are NEVER\nreturned inline.",
        "properties": {
          "name": {
            "title": "Name",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "format": {
            "title": "Format",
            "type": "string"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "checksum": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checksum"
          },
          "trace": {
            "additionalProperties": true,
            "title": "Trace",
            "type": "object"
          }
        },
        "required": [
          "name",
          "version",
          "format"
        ],
        "title": "ResolvedItemSchema",
        "type": "object"
      },
      "ResolveParametersResponseSchema": {
        "properties": {
          "path": {
            "title": "Path",
            "type": "string"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "requested_version": {
            "title": "Requested Version",
            "type": "string"
          },
          "parameters": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ResolvedParameterSchema"
            },
            "title": "Parameters",
            "type": "object"
          }
        },
        "required": [
          "path",
          "version",
          "requested_version"
        ],
        "title": "ResolveParametersResponseSchema",
        "type": "object"
      },
      "ResolvedParameterSchema": {
        "description": "One parameter row in the `/~resolve-parameters/` debug response.",
        "properties": {
          "type": {
            "title": "Type",
            "type": "string"
          },
          "description": {
            "default": "",
            "title": "Description",
            "type": "string"
          },
          "selector": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Selector"
          },
          "secret_reference": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Secret Reference"
          },
          "declared_default": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Declared Default"
          },
          "raw_value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Raw Value"
          },
          "effective_value": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Effective Value"
          },
          "transformers_applied": {
            "items": {
              "type": "string"
            },
            "title": "Transformers Applied",
            "type": "array"
          },
          "caller_supplied": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Caller Supplied"
          }
        },
        "required": [
          "type"
        ],
        "title": "ResolvedParameterSchema",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "ResolverAuth": {
        "type": "apiKey",
        "in": "query",
        "name": "token"
      },
      "OAuth2AuthorizationCodeBearer": {
        "type": "oauth2",
        "flows": {
          "authorizationCode": {
            "scopes": {
              "openid": "",
              "profile": "",
              "email": "",
              "groups": ""
            },
            "authorizationUrl": "http://localhost:8080/dex/auth",
            "tokenUrl": "http://localhost:8080/dex/token"
          }
        }
      }
    }
  },
  "servers": []
}